VISTA-56 #time 3h do not send byte[] picture to GWT, use base64; other serialization fixes

+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/Money.java
@@ -18,7 +18,7 @@ import com.pyx4j.entity.shared.IPrimitive;
 
 public interface Money extends IEntity {
 
-    IPrimitive<Float> amount();
+    IPrimitive<Double> amount();
 
     Currency currency();
 }
